developing a faster scraper
