+++ eventbrite_scraper/utils.py
@@ -1,0 +1,10 @@
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKCYAN = '\033[96m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ESCAPE = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
