+++ scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = event_scraper.settings
+default = eventbrite_scraper.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = event_scraper
+project = eventbrite_scraper
